fix: mv PyannoteDetectorArgs to types/detector
