+++ src/common/types.py
@@ -200,22 +200,6 @@ class WebRTCSileroVADArgs(WebRTCVADArgs, SileroVADArgs):
     pass
 
 
-@dataclass
-class PyannoteDetectorArgs:
-    from pyannote.audio.core.io import AudioFile
-    hf_auth_token: str = ""  # defualt use env HF_TOKEN
-    path_or_hf_repo: str = "pyannote/segmentation-3.0"
-    model_type: str = "segmentation-3.0"
-    # remove speech regions shorter than that many seconds.
-    min_duration_on: float = 0.3
-    # fill non-speech regions shorter than that many seconds.
-    min_duration_off: float = 0.3
-    # if use pyannote/segmentation open onset/offset activation thresholds
-    onset: float = 0.5
-    offset: float = 0.5
-    # vad
-    vad_pyannote_audio: AudioFile = None
-
 
 INIT_WAKE_WORDS_SENSITIVITY = 0.6
 INIT_WAKE_WORD_ACTIVATION_DELAY = 0.0
