+++ src/types/speech/detector/pyannote.py
@@ -1,0 +1,19 @@
+from dataclasses import dataclass
+
+from pyannote.audio.core.io import AudioFile
+
+
+@dataclass
+class PyannoteDetectorArgs:
+    hf_auth_token: str = ""  # defualt use env HF_TOKEN
+    path_or_hf_repo: str = "pyannote/segmentation-3.0"
+    model_type: str = "segmentation-3.0"
+    # remove speech regions shorter than that many seconds.
+    min_duration_on: float = 0.3
+    # fill non-speech regions shorter than that many seconds.
+    min_duration_off: float = 0.3
+    # if use pyannote/segmentation open onset/offset activation thresholds
+    onset: float = 0.5
+    offset: float = 0.5
+    # vad
+    vad_pyannote_audio: AudioFile = None
